+++ examples/example_typical_usage.py
@@ -1,4 +1,5 @@
-from glossarpy import glossarpy
+from glossarpy import *
+
 '''
 glossarpy typical use case example
