--- examples/imported_entries_output.rst
@@ -1,35 +1,0 @@
-.. This file was created using GreatGloss. It is highly recommended to update the
-        source file that this page was generated from rather than modifying it directly.
-
-They're Good Dogs, Brent
-========================
-.. hlist:: 
-	:columns: 3
-
-	* :ref:`dict canine`
-	* :ref:`dict wolf`
-
-.. _dict canine:
-
-canine
-------
-    A digitigrade, mostly-carnivorous animal in the Canidae family of mammals  
-
-Further reading: `<https://en.wikipedia.org/wiki/Canidae>`_  
-
-.. updated 2022-05-17  
-
-
-
-.. _dict wolf:
-
-wolf
-----
-    A large :ref:`dict canine` found across the Northern Hemisphere known to form packs  
-
-Further reading: `<https://en.wikipedia.org/wiki/Wolf>`_  
-
-.. updated 2022-05-17  
-
-
-
