--- examples/typical_usage_output.rst
@@ -1,59 +1,0 @@
-.. This file was created using GreatGloss. It is highly recommended to update the
-        source file that this page was generated from rather than modifying it directly.
-
-Very Cool Animals
-=================
-.. hlist:: 
-	:columns: 3
-
-	* :ref:`dict cat`
-	* :ref:`dict Juice`
-	* :ref:`dict Roofcat`
-	* :ref:`dict tiger`
-
-.. _dict cat:
-
-cat
----
-    A digitigrade carnivorous animal in the Felidae family of mammals  
-
-Further reading: `<https://en.wikipedia.org/wiki/Felidae>`_  
-
-.. updated 2022-05-17  
-
-
-
-.. _dict Juice:
-
-Juice
------
-    A mysterious orange housecat, also known as Roofcat, native to the East Side of Santa Cruz who successfully charmed at least four different households into feeding him at the same time  
-
-Further reading: `<https://roofcat.care>`_  
-
-.. updated 2022-05-17  
-
-
-
-.. _dict Roofcat:
-
-Roofcat
--------
-see also :ref:`dict Juice`  
-
-.. updated 2022-05-17  
-
-
-
-.. _dict tiger:
-
-tiger
------
-    An orange-and-black striped :ref:`dict cat` native to Southeast Asia  
-
-Further reading: `<https://en.wikipedia.org/wiki/Tiger>`_  
-
-.. updated 2022-05-17  
-
-
-
