--- examples/typical_usage_toc.txt
@@ -1,4 +1,0 @@
-cat
-Juice
-Roofcat
-tiger
